refactor: 붎필요한 파일 제거
