--- src/modules/CartItemQuery/useCartItemsQuery.ts
@@ -1,9 +1,0 @@
-import { CartItemApi } from "@/apis";
-import { useQuery } from "@/modules";
-
-export function useCartItemsQuery() {
-  return useQuery({
-    queryFn: CartItemApi.getCartItems,
-    queryKey: "cartItems",
-  });
-}
