--- src/modules/CartItemQuery/usePatchCartItemMutation.ts
@@ -1,8 +1,0 @@
-import { CartItemApi, PatchCartItemsParams } from "@/apis";
-import { useMutation } from "@/modules";
-
-export function usePatchCartItemMutation() {
-  return useMutation<PatchCartItemsParams, void>({
-    mutationFn: CartItemApi.patchCartItems,
-  });
-}
