+++ src/pages/products/hooks/index.ts
@@ -1,1 +1,1 @@
-export { default as useCartItem } from "./useCartItem";
+export { default as useCartItem } from "./useCartItem/useCartItem";
